Extract component renderers for agent and agent details
